Don't set the port, just let it default.
It will default to port 80 in Meteor.  This maintains compatibility with all existing Mup, and meteord.  Additionally, I don't believe there is any reason to not run as port 80, inside the Docker container.  This is NOT the exposed port (which you would set with the `-p WANT_PORT:80` argument to `docker run`) and we're permitted to run, unprivileged, as port 80 inside a container (thanks to netcaps by the Docker engine).  It is no more or less secure than using port 3000 in this environment.

While this was previously "configurable", I don't think it has ever worked because the Dockerfile has only ever EXPOSE-ed port 80.

Also changed tests to match.