Use port 3000 until docker really supports NET_BIND_SERVICE
For an unfortunate variety of reasons (mainly portability issues dealing with Docker kernel caps, Debian vs Mac, CircleCI vs others, AUFS vs blah, etc., etc., etc.) we cannot currently rely on the NET_BIND_SERVICE being set, even though it's supposed to be on by default (https://docs.docker.com/engine/reference/run/#/runtime-privilege-and-linux-capabilities).  This seems like it will be fixed in Docker 1.14, but until then we'll have to resort to just running on port 3000.

Unfortunately, this means that we are helpless to solve the meteor-up problem in which it _expects_ the image to be listening on port 80.  A PR will have to be made to `mup` to fix this.

I tried. :(